+++ idea/testData/quickfix/createFromUsage/createFunction/binaryOperations/typeMismatch.kt
@@ -1,0 +1,8 @@
+// "Create extension function 'A.times'" "true"
+class A
+
+operator fun A.times(i: Int) = this
+
+fun test() {
+    A() * <caret>"1"
+}